Create BEAMweatherLightSensors.py
This file contains the code to run the BME-280 weather sensor and the TSL-2561 light sensor through the use of i2c and SPI communication, respectively, with a Raspberry Pi. After data is collected from the sensors, it is sent (appended) to a JSON file also found on the Pi.